requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-filter==24.3
qrcode==7.4.2
Pillow==10.4.0
python-dotenv==1.0.1
